+++ src/parkapi_sources/converters/friedrichshafen_sensors/converter.py
@@ -6,11 +6,11 @@ Use of this source code is governed by an MIT-style license that can be found in
 from validataclass.exceptions import ValidationError
 from validataclass.validators import DataclassValidator
 
-from parkapi_sources.converters.base_converter import ParkingSpotBaseConverter
 from parkapi_sources.converters.base_converter.datex2 import ParkingRecordStatusMixin, UrbanParkingSiteMixin
-from parkapi_sources.converters.base_converter.pull import MobilithekPullConverter
+from parkapi_sources.converters.base_converter.pull import MobilithekPullConverterMixin, ParkingSpotPullConverter
 from parkapi_sources.exceptions import ImportParkingSpotException
 from parkapi_sources.models import RealtimeParkingSpotInput, SourceInfo, StaticParkingSpotInput
+from parkapi_sources.models.enums import ParkingSpotType
 
 from .validators import FriedrichshafenSensorsParkingRecordStatus, FriedrichshafenSensorsParkingSpot
 
@@ -18,8 +18,8 @@ from .validators import FriedrichshafenSensorsParkingRecordStatus, Friedrichshaf
 class FriedrichshafenSensorsPullConverter(
     UrbanParkingSiteMixin,
     ParkingRecordStatusMixin,
-    ParkingSpotBaseConverter,
-    MobilithekPullConverter,
+    MobilithekPullConverterMixin,
+    ParkingSpotPullConverter,
 ):
     config_key = 'FRIEDRICHSHAFEN_SENSORS'
     static_validator = DataclassValidator(FriedrichshafenSensorsParkingSpot)
@@ -46,6 +46,7 @@ class FriedrichshafenSensorsPullConverter(
             try:
                 static_item = self.static_validator.validate(static_input_dict)
                 static_parking_spot_input = static_item.to_static_parking_spot_input()
+                static_parking_spot_input.type = ParkingSpotType.ON_STREET
 
                 static_parking_spot_inputs.append(static_parking_spot_input)