+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # Changelog
 
+## 0.18.2
+
+Released 2025-05-10
+
+### Fixes
+
+* [better datex2 handling and testing](https://github.com/ParkenDD/parkapi-sources-v3/pull/212)
+
+
+## 0.18.1
+
+Released 2025-05-08
+
+### Fixes
+
+* [better handling of mobilithek pull mechanisms](https://github.com/ParkenDD/parkapi-sources-v3/pull/211)
+
+
 ## 0.18.0
+
+Released 2025-04-02
 
 This release is a major change, as it introduces parking spots. If you used internal objects, this update might be
 breaking for you. Existing converters will work exactly the same and have the same interface, so for most users,
@@ -14,6 +34,7 @@ updating will be fine.
 
 ## 0.17.4
 
+Released 2025-04-01
 
 ### Fixes